Added Metamodel suppport and Proxies
The FEM now is capable of generating the source code for the
static metamodel classes and automcatically populates them with
attribute values when the entity manager factory is set to 
'manage(...)' the classes

A simple tool including GUI interface is available for automatically
generating static metamodel classes.

Lazy loaded links are populated with proxy classes to fetch the linked
entity lazily.